[#46] Rename: 도메인 하위 패키지명 변경
