+++ src/main/java/com/server/money_touch/domain/budget/repository/budgetCategory/BudgetCategoryRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.server.money_touch.domain.budget.repository.budget_category;
+package com.server.money_touch.domain.budget.repository.budgetCategory;
 
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
