+++ src/main/java/com/server/money_touch/domain/budget/service/budget/BudgetQueryServiceImpl.java
@@ -6,10 +6,10 @@ import com.server.money_touch.domain.budget.entity.Budget;
 import com.server.money_touch.domain.budget.entity.BudgetCategory;
 import com.server.money_touch.domain.budget.enums.CategoryType;
 import com.server.money_touch.domain.budget.repository.budget.BudgetRepository;
-import com.server.money_touch.domain.budget.repository.budget_category.BudgetCategoryRepository;
-import com.server.money_touch.domain.consumptionRecord.converter.total_consumption.TotalConsumptionConverter;
+import com.server.money_touch.domain.budget.repository.budgetCategory.BudgetCategoryRepository;
+import com.server.money_touch.domain.consumptionRecord.converter.totalConsumption.TotalConsumptionConverter;
 import com.server.money_touch.domain.consumptionRecord.entity.TotalConsumption;
-import com.server.money_touch.domain.consumptionRecord.repository.total_consumption.TotalConsumptionRepository;
+import com.server.money_touch.domain.consumptionRecord.repository.totalConsumption.TotalConsumptionRepository;
 import com.server.money_touch.domain.user.entity.User;
 import com.server.money_touch.domain.user.respotiroy.user.UserRepository;
 import com.server.money_touch.global.apiPayload.code.status.ErrorStatus;
