+++ src/main/java/com/server/money_touch/domain/consumptionRecord/converter/totalConsumption/TotalConsumptionConverter.java
@@ -1,4 +1,4 @@
-package com.server.money_touch.domain.consumptionRecord.converter.total_consumption;
+package com.server.money_touch.domain.consumptionRecord.converter.totalConsumption;
 
 import com.server.money_touch.domain.consumptionRecord.entity.TotalConsumption;
 import com.server.money_touch.domain.user.entity.User;
